requirements.txt
sentence-transformers>=2.7.0
python-frontmatter>=1.1.0
